renamed UI and Locator modules to avoid wrong referencing
