+++ pyChilizer.tab/Project.panel/rooms.stack/Room Data Sheets.pulldown/Room Data Sheets.pushbutton/script.py
@@ -1,6 +1,6 @@
 from pyrevit import revit, DB, script, forms
 from rpw.ui.forms import FlexForm, Label, TextBox, Button, ComboBox, CheckBox, Separator
-import locator, ui
+import rdslocator, rdsui
 from itertools import izip
 import sys
 from pychilizer import units, select, geo, database
@@ -11,7 +11,7 @@ from Autodesk.Revit import Exceptions
 
 
 
-ui = ui.UI(script)
+ui = rdsui.UI(script)
 ui.is_metric = units.is_metric
 doc = __revit__.ActiveUIDocument.Document
 
@@ -250,7 +250,7 @@ for room in selection:
         sheet = database.create_sheet(chosen_sheet_nr, room_name_nr, chosen_tb.Id)
 
     # get positions on sheet
-    loc = locator.Locator(sheet, titleblock_offset, tb_ori, layout_ori)
+    loc = rdslocator.Locator(sheet, titleblock_offset, tb_ori, layout_ori)
     plan_position = loc.plan
     RCP_position = loc.rcp
     elev_positions = loc.elevations
